using sharding-transaction, refactor proxy and jdbc.
