+++ sharding-jdbc/pom.xml
@@ -14,6 +14,11 @@
             <artifactId>sharding-core</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>io.shardingsphere</groupId>
+            <artifactId>sharding-transaction</artifactId>
+            <version>${project.version}</version>
+        </dependency>
         
         <dependency>
             <groupId>org.projectlombok</groupId>
