+++ sharding-transaction/src/main/java/io/shardingsphere/transaction/common/event/WeakXaTransactionEvent.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.core.transaction.event;
+package io.shardingsphere.transaction.common.event;
 
 import com.google.common.base.Optional;
 import io.shardingsphere.core.constant.TCLType;
