+++ sharding-transaction/src/main/java/io/shardingsphere/transaction/common/spi/TransactionManager.java
@@ -15,10 +15,9 @@
  * </p>
  */
 
-package io.shardingsphere.core.transaction.spi;
+package io.shardingsphere.transaction.common.spi;
 
-import io.shardingsphere.core.transaction.event.TransactionEvent;
-
+import io.shardingsphere.transaction.common.event.TransactionEvent;
 import java.sql.SQLException;
 
 /**
