+++ sharding-transaction/src/main/java/io/shardingsphere/transaction/weakxa/WeakXaTransactionManager.java
@@ -15,11 +15,11 @@
  * </p>
  */
 
-package io.shardingsphere.transaction;
+package io.shardingsphere.transaction.weakxa;
 
-import io.shardingsphere.core.transaction.event.TransactionEvent;
-import io.shardingsphere.core.transaction.event.WeakXaTransactionEvent;
-import io.shardingsphere.core.transaction.spi.TransactionManager;
+import io.shardingsphere.transaction.common.event.TransactionEvent;
+import io.shardingsphere.transaction.common.event.WeakXaTransactionEvent;
+import io.shardingsphere.transaction.common.spi.TransactionManager;
 
 import java.sql.Connection;
 import java.sql.SQLException;
